feat: refactoring of the User SummaryReport.
